requirements.txt
argtools==0.1.6
biopython==1.79
Cython==0.29.33
DendroPy==4.5.2
deepdiff==6.3.0
ete3==3.1.2
numpy==1.22.4
omegaconf==2.3.0
pandas==1.2.4
PyYAML==5.3.1
scipy==1.6.3
seaborn==0.11.1
torch==1.13.1
tqdm==4.60.0
